refactor: delete ferrum-runtime crate, move CandleBackend into ferrum-engine (Phase 5b/2)
Closes the second half of Phase 5b: removes the legacy ferrum-runtime
crate and folds its still-used pieces into ferrum-engine.

What moves (rename, no content change):
- ferrum-runtime/src/backends/candle.rs           → ferrum-engine/src/backends/candle.rs
- ferrum-runtime/src/backends/candle_kernel_ops.rs → ferrum-engine/src/backends/candle_kernel_ops.rs
- ferrum-runtime/src/memory/{mod,pool,stats}.rs    → ferrum-engine/src/memory/

What gets dropped (no consumers in the new architecture):
- ferrum-runtime/src/backends/cpu.rs           (legacy CPU `ComputeBackend` impl, never instantiated)
- ferrum-runtime/src/backends/mod.rs           (replaced by a slim mod.rs in ferrum-engine)
- ferrum-runtime/src/lib.rs                    (DefaultBackendRegistry / global_backend_registry — never consumed)
- ferrum-runtime/Cargo.toml                    (~3 KB)

Import rewires:
- `ferrum_runtime::backends::candle::CandleTensorFactory` → `crate::backends::candle::CandleTensorFactory`
  (call site: ferrum-engine/src/builder.rs:387)
- `ferrum_runtime::backends::CandleBackend`               → `crate::backends::candle::CandleBackend`
  (call sites: ferrum-engine/src/registry.rs:481, 873)
- `ferrum_runtime::backends::CandleTensor`                → `crate::backends::candle::CandleTensor`
  (call site: ferrum-engine/src/metal/metal_kernel_ops.rs:17)
- `pub use ferrum_runtime::ComputeBackend`                → `pub use ferrum_interfaces::ComputeBackend`
  (re-export hub: ferrum-engine/src/lib.rs)
- candle.rs internal imports: `use crate::{ComputeBackend, MemoryPool, …}` → split into
  `use crate::memory::MemoryPool` + `use ferrum_interfaces::{ComputeBackend, …}` since
  ferrum-runtime's lib.rs re-exports are gone.

Cargo.toml diet:
- ferrum-engine: drop `ferrum-runtime`, add the three deps the moved files need
  (`once_cell`, `crossbeam-channel`, `crossbeam-utils`, `pin-project`, `bytesize`).
- ferrum-cli, ferrum-kv, ferrum-models: drop the `ferrum-runtime` dep entirely
  (they only listed it in Cargo.toml; no `use ferrum_runtime` anywhere in their src).
- workspace Cargo.toml: drop `crates/ferrum-runtime` from members + the workspace dep entry.

Net effect:
- 1 crate gone (`ferrum-runtime`).
- ~600 lines deleted (cpu.rs / lib.rs / Cargo.toml / mod.rs scaffolding).
- ferrum-engine now self-contains the only `ComputeBackend` impl that anybody
  actually instantiates (the registry's stub-executor factory uses it).
- New backends should target `ferrum_kernels::backend::Backend<B>` (the
  production trait), not `ComputeBackend`. The latter survives only as a
  fallback abstraction for the stub executor + registry plumbing.

Acceptance:
  cargo check --workspace --all-targets        clean
  cargo test --workspace --features metal --lib   clean (all 14 crates pass)
  cargo fmt --all -- --check                   clean
