+++ crates/ferrum-engine/src/builder.rs
@@ -382,10 +382,9 @@ impl EngineBuilder {
         );
 
         // Create TensorFactory for the configured device
-        let tensor_factory: Arc<dyn TensorFactory> =
-            Arc::new(ferrum_runtime::backends::candle::CandleTensorFactory::new(
-                config.backend.device.clone(),
-            ));
+        let tensor_factory: Arc<dyn TensorFactory> = Arc::new(
+            crate::backends::candle::CandleTensorFactory::new(config.backend.device.clone()),
+        );
 
         // Opt-in speculative decoding: set FERRUM_SPEC_DRAFT=<model_path>
         // (absolute path to a HF snapshot dir) to load a second smaller
