+++ crates/ferrum-engine/src/memory/stats.rs
