--- crates/ferrum-runtime/Cargo.toml
@@ -1,63 +1,0 @@
-[package]
-name = "ferrum-runtime"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-license.workspace = true
-description = "Backend implementations (Candle, CPU) for Ferrum inference"
-readme = "../../README.md"
-
-[dependencies]
-# Core dependencies - New architecture
-ferrum-types = { workspace = true }
-ferrum-interfaces = { workspace = true }
-
-# Async runtime
-async-trait = { workspace = true }
-tokio = { workspace = true }
-futures = { workspace = true }
-parking_lot = { workspace = true }
-
-# Serialization
-serde = { workspace = true }
-serde_json = { workspace = true }
-
-# Error handling
-anyhow = { workspace = true }
-thiserror = { workspace = true }
-
-# Logging and metrics
-tracing = { workspace = true }
-metrics = { workspace = true }
-
-# Tensor operations - Candle backend
-candle-core = { workspace = true, optional = true }
-candle-nn = { workspace = true, optional = true }
-
-# Async utilities
-crossbeam-channel = "0.5"
-crossbeam-utils = "0.8"
-pin-project = "1.1"
-
-# Utils
-uuid = { workspace = true }
-chrono = { workspace = true }
-bytesize = "1.3"
-once_cell = { workspace = true }
-
-# Memory management
-ndarray = { workspace = true, optional = true }
-half = { workspace = true }
-
-[dev-dependencies]
-tokio-test = "0.4"
-criterion = "0.5"
-
-[features]
-default = ["candle"]
-candle = ["dep:candle-core", "dep:candle-nn"]  # Candle backend
-cpu = ["dep:ndarray"]                          # CPU backend
-metal = []                                     # Metal runtime support
-cuda = []                                      # CUDA runtime support
-rocm = []                                      # ROCm runtime support
-profiling = []                                 # Enable detailed profiling
